move function from args.py to frame_interp file
